+++ scripts/DBUpdateTo6/MigrateTicketStorageModule.pm
@@ -28,6 +28,14 @@ scripts::DBUpdateTo6::MigrateTicketStorageModule - Migrate Ticket::StorageModule
 sub Run {
     my ( $Self, %Param ) = @_;
 
+    my $Home     = $Kernel::OM->Get('Kernel::Config')->Get('Home');
+    my $FilePath = "$Home/Kernel/Config/Backups/ZZZAutoOTRS5.pm";
+
+    if ( !-f $FilePath ) {
+        print "\nCould not find Kernel/Config/Backups/ZZZAutoOTRS5.pm, skipping... ";
+        return 1;
+    }
+
     my $SysConfigObject = $Kernel::OM->Get('Kernel::System::SysConfig');
 
     my %OTRS5Config;